requirements.txt
fastapi
python-multipart
starlette
uvicorn
pillow
pytesseract
python-dotenv